1. es 版本调整 7.15.0 2. 交易统计分析结果查询 数据合并 3. 交易统计分析结果查询 数据内存排序与分页 4. 交易统计分析结果查询响应体 对应自定义注解定义

+++ tldw-aggregate-searching/tldw-aggregate-commons/src/main/java/com/zqykj/domain/PageRequest.java
@@ -84,7 +84,7 @@ public class PageRequest extends AbstractPageRequest {
     }
 
     // 计算总页数
-    public static int getTotalPages(int total, int pageSize) {
+    public static int getTotalPages(long total, int pageSize) {
         return total == 0 ? 0 : (int) Math.ceil((double) total / (double) pageSize);
     }
 }
